Provide `lief-patchelf`: a LIEF-based implementation of NixOS's patchelf
Related: NixOS/patchelf#568
Support: NixOS/patchelf#596

+++ api/rust/cargo/lief/src/elf/segment.rs
@@ -180,7 +180,7 @@ impl Flags {
     }
 
     pub fn to_value(&self) -> u32 {
-        self.bits() as u32
+        self.bits()
     }
 }
